refactor: move patient info form to component
