+++ pages/patient-info.tsx
@@ -1,71 +1,13 @@
-import { z } from 'Zod'
-import { zodResolver } from '@hookform/resolvers/zod'
-import { useMutation, useQuery } from '@tanstack/react-query'
-import { SubmitHandler, useForm } from 'react-hook-form'
+import { useQuery } from '@tanstack/react-query'
 import { MainLayout } from '../components/layouts/mainLayout'
 import { BackButton } from '../components/shared/BackButton'
-import { BaseInput } from '../components/shared/BaseInput'
-import { BaseTextArea } from '../components/shared/BaseTextArea'
-import { BaseButton } from '../components/shared/BaseButton'
-import { postPatientInfo } from '../api/patient'
-import toast, { Toaster } from 'react-hot-toast'
 import { getUser } from '../api/user'
-import { useEffect } from 'react'
-
-const patientInfoSchema = z.object({
-  phone: z.string().min(1, { message: 'Phone number is required' }),
-  weight: z.string().min(1, { message: 'Weight is required' }),
-  height: z.string().min(1, { message: 'Height is required' }),
-  info: z.string(),
-})
-
-export type FormValues = z.infer<typeof patientInfoSchema>
+import { PatienInfoForm } from '../components/patientInfo/PatientInfoForm'
+import { Oval } from 'react-loader-spinner'
 
 export default function PatientInfoPage() {
-  const { data: userData } = useQuery(['getUser'], getUser)
+  const { data: userData, isLoading } = useQuery(['getUser'], getUser)
 
-  const {
-    formState: { errors },
-    register,
-    handleSubmit,
-    reset
-  } = useForm<FormValues>({ resolver: zodResolver(patientInfoSchema), defaultValues: {
-    phone: userData?.info?.phone,
-    weight: userData?.info?.weight,
-    height: userData?.info?.height,
-    info: userData?.info?.info
-  }})
-
-  const getData = (data: FormValues) => {
-    const newData = {
-      phone: data.phone,
-      weight: data.weight,
-      height: data.height,
-      info: data.info ? data.info : 'No comments',
-    }
-    return newData
-  }
-  const { mutate } = useMutation({
-    mutationFn: postPatientInfo,
-    onSuccess: () => {
-      toast.success('Information successfully updated', {
-        position: 'top-right',
-      })
-    },
-    onError: (error: any) => {
-      toast.error(error.response.data.message, {
-        position: 'top-right',
-      })
-    },
-  })
-
-  const onSubmit: SubmitHandler<FormValues> = data => {
-    mutate(getData(data))
-  }
-
-  useEffect(() => {
-    reset()
-  }, [userData])
   return (
     <MainLayout userType="PATIENT">
       <div className="px-3">
@@ -77,48 +19,26 @@ export default function PatientInfoPage() {
           </h1>
         </div>
       </div>
-      <form className="w-[513px] mt-4" onSubmit={handleSubmit(onSubmit)}>
-        <div className="w-full flex flex-row items-center justify-between">
-          <BaseInput
-            label="Phone number"
-            type="number"
-            error={errors.phone}
-            errorMessage={errors.phone?.message}
-            {...register('phone')}
-          />
-        </div>
-        <div className="mt-1 w-full flex flex-row items-center justify-between">
-          <BaseInput
-            label="Weight"
-            type="number"
-            error={errors.weight}
-            errorMessage={errors.weight?.message}
-            {...register('weight')}
-          />
-          <BaseInput
-            label="Height"
-            type="number"
-            error={errors.height}
-            errorMessage={errors.height?.message}
-            {...register('height')}
-          />
-        </div>
-        <div className="mt-1 w-full flex flex-row items-center justify-between px-3">
-          <BaseTextArea
-            {...register('info')}
-            label="Other info"
-            rows={7}
-          />
-        </div>
-        <div className="mt-6 w-1/3 px-3">
-          <BaseButton
-            buttonClass="primary"
-            text="Update profile"
-            type="submit"
-          />
-        </div>
-      </form>
-      <Toaster />
+      {userData ? (
+        <PatienInfoForm userData={userData} />
+      ) : (
+        <Oval
+          height={20}
+          width={20}
+          color="#2563EB"
+          wrapperStyle={{
+            position: 'absolute',
+            right: '50%',
+            top: '50%',
+          }}
+          wrapperClass=""
+          visible={isLoading}
+          ariaLabel="oval-loading"
+          secondaryColor="#E0F2FE"
+          strokeWidth={2}
+          strokeWidthSecondary={2}
+        />
+      )}
     </MainLayout>
   )
 }
